Updated DB module
