+++ examplecore/src/main/scala/examples/Job4.scala
@@ -12,11 +12,13 @@ object Job4 extends JobApp {
 
   val cred = JDBC(sys.env("LOG_DB_URL"), sys.env("LOG_DB_USER"), sys.env("LOG_DB_PWD"), "org.postgresql.Driver")
 
+  // override val log_layer: ZLayer[zio.ZEnv, Throwable, LogEnv] = etlflow.log.DBLogger(cred)
+
   private def step1(cred: JDBC): DBReadStep[EtlJobRun] = DBReadStep[EtlJobRun](
     name = "FetchEtlJobRun",
-    query = "SELECT job_name,job_run_id,state FROM jobrun LIMIT 10",
+    query = "SELECT job_name,job_run_id,status FROM jobrun LIMIT 10",
     credentials = cred
-  )(rs => EtlJobRun(rs.string("job_name"), rs.string("job_run_id"), rs.string("state")))
+  )(rs => EtlJobRun(rs.string("job_name"), rs.string("job_run_id"), rs.string("status")))
 
   private def processData(ip: List[EtlJobRun]): Unit = {
     logger.info("Processing Data")
