+++ modules/db/src/main/scala/etlflow/db/package.scala
@@ -78,11 +78,11 @@ package object db {
   }
 
   def liveDB(db: JDBC, pool_name: String = "EtlFlow-Pool", pool_size: Int = 2): ZLayer[Blocking, Throwable, DBEnv] =
-    Implementation.cpLayer(db, pool_name, pool_size) >>> Implementation.dbLayer
+    CP.layer(db, pool_name, pool_size) >>> DBImpl.live
 
   private[etlflow] def liveLogServerDB(db: JDBC, pool_name: String = "EtlFlow-Pool", pool_size: Int = 2): ZLayer[Blocking, Throwable, LogEnv with DBServerEnv] =
-    Implementation.cpLayer(db, pool_name, pool_size) >>> (log.DBLogger.dbLogLayer ++ Implementation.dbServerLayer)
+    CP.layer(db, pool_name, pool_size) >>> (log.DBLogger.live ++ DBServerImpl.live)
 
   private[etlflow] def liveFullDB(db: JDBC, pool_name: String = "EtlFlow-Pool", pool_size: Int = 2): ZLayer[Blocking, Throwable, DBEnv with LogEnv with DBServerEnv] =
-    Implementation.cpLayer(db, pool_name, pool_size) >>> (Implementation.dbLayer ++ log.DBLogger.dbLogLayer ++ Implementation.dbServerLayer)
+    CP.layer(db, pool_name, pool_size) >>> (DBImpl.live ++ log.DBLogger.live ++ DBServerImpl.live)
 }
