+++ modules/db/src/test/scala/etlflow/db/UtilsTestSuite.scala
@@ -1,9 +1,7 @@
 package etlflow.db
 
-import etlflow.db.Sql.getStartTime
 import zio.test.Assertion.equalTo
 import zio.test._
-
 import java.text.SimpleDateFormat
 import java.time.LocalDate
 
@@ -12,7 +10,7 @@ object UtilsTestSuite{
   val spec: ZSpec[environment.TestEnvironment, Any] = {
     suite("UtilsTestSuite Api")(
       test("Provided Exception should return") {
-        assert(getStartTime(None))(equalTo(sdf.parse(LocalDate.now().toString).getTime))
+        assert(Utils.getStartTime(None))(equalTo(sdf.parse(LocalDate.now().toString).getTime))
       }
     )
   }
